Create new BlockDevBackingStore abstraction.
Separate ideas of BlockDev and BlockDevBackingStore:

- BlockDev is the API of a virtualized block device.
- BlockDevBackingStore fulfils the requests sent to the BlockDev.

PlainBdev does the same thing that it did before, except now requests are
fulfilled by a supplied BlockDevBackingStore. New backing stores now only need
to implement a new BlockDevBackingStore and hand it to a PlainBDev. Parts of
PlainBdev that interacted with a file have been pulled out into a new
FileBlockDevBackingStore.

Functionally, this commit only represents a refactor. The next one will
implement a BlockDevBackingStore for Crucible.