nvme: zero out offset of prp2 if in prp list
prp2 points to a prp list if the size of the operation is large enough.

In the 1.0e spec, prp2 "may also have a non-zero offset within the
memory page", but it also says "If this entry is in the PRP List then
the Offset portion of this field shall be cleared to 0h".

This commit clears the offset, fixing errors seen when using NVMe
emulation.