Accept crucible disks in instance ensure request
Add disks to the InstanceEnsureRequest, and attach as crucible disks.
Disks that are specified in the server_context.config still get
attached.

Fill in the logic for SlotType::disk to make slot_to_bdf work.

Add a tokio runtime as part of the InstanceContext, making sure it lives
for the lifetime of the instance itself.

Adds slog logging.