Add Crucible as a block backend (#56)
Implement block::Backend for new CrucibleBackend type. Copy Driver from
the FileBackend code, with the intention in the future of submitting
work asynchronously.

Add Crucible as optional - it's not necessary to compile with it.

Note dispatcher spawned task names have to be globally unique! Added
a print of {:?} to FileBackend code.

BlockReqWaiter is stalling tokio executor threads, so use block_in_place

spawn_blocking is not applicable here because CrucibleBackend::create isn't async