Accept migration UUID as part of API request.
Instead of generating the common migration UUID used across both the
source and destination propolis instances, we have the initiator pass it
in. This lets us thread the same ID used as part of the nexus migration
saga to the destination sled agent and then on to the individual
propolis instances. This should allow us to easily correlate migration
events and logs across the different moving parts involved.