fix break that will be caused by merging #182
Crucible PR https://github.com/oxidecomputer/crucible/pull/182 will
break compilation for propolis because new CrucibleOpts fields were
added there. Add the Default line here so that this doesn't happen in
the future, but Default was also derived for CrucibleOpts in that PR.
Unfortunately, these have to go in at the same time.