Async access to machine state during migration (#89)
This adds a new 'Migrating' state as part of the internal propolis state machine [1]. As part of transitioning to that state, we allow specifying a `CtxId` representing an `AsyncCtx` object associated with an async task. This lets us quiesce any other tasks but explicitly allow the migrate task to access the Instance's `DispCtx` to obtain the underlying `MachineCtx` and vmm handles.