Provide direct access to memory segments
Memory objects (segments in the bhyve terminology) are mapped into the
guest-physical address space, sometimes with restricted permissions.
There are times when direct access to the underlying objects is
necessary, like when loading the bootrom (which is mapped read-only) or
migrating guest RAM (when the accessed/dirty bit changes tracked by the
nested page tables are to be circumvented).