Implement "NoCloud" cloud-init
Accept new CloudInit field as part of instance ensure request. The user
can specify meta-data, and either structured fields (which renders
to #cloud-config), or a custom base64 encoded string that renders to
user-data.

Build a CIDATA ISO from these two files and mount the bytes it as a
virtio device. This commit also adds a new in-memory block backend for
propolis.