Threading migration through instance state machine. (#92)
We currently have the migration being driven from `propolis-server` but it needs to be able inspect and influence the `Instance` in the core propolis lib. `Instance` already had a 'migrating' state but this expands it with some phases: (just `Start` and `Pause` for now). A migration begins in the `Start` phase.

Once the migrate task is ready to pause the source instance it requests a transition to the `Pause` phase. But pausing the VM involves having every device stop servicing the guest, along with making sure any outstanding requests are completed (or canceled). To support that, the `Migrate(Source, Pause)` state triggers the normal state transition logic which notifies each device in turn (via `Entity::state_transition` along with a shortcut `Migrate::pause`). This is the point at which each device is expected to stop servicing guest requests.

Crucially though, we can't rely on every device being ready to pause at that moment due to the possibility of outstanding requests. To handle that, every device also implements `Migrate::paused` which returns a future that indicates the device is done pausing.

Right now, the state machine just waits on a passed in channel to allow `propolis-server` to actually collect and await on the device futures.