Use less IDs in paths (#117)
## Old Behavior

For most endpoints, we'd use the format: 

```
http://<IP ADDRESS>/instances/<INSTANCE UUID>/<rest of the path, maybe>
```

For many endpoints, this UUID would be redundant with values stored in `properties` struct, and a somewhat silly check would validate this. This path format added little value, other than requiring clients to supply a redundant value.

Why is the UUID redundant? Propolis' server only supports a *single* instance anyway, so the server address should be a sufficient factor for distinguishing instances.

## New Behavior

Don't bother supplying the instance UUID. Simplify a lot of "name -> UUID" pathways where this is no longer necessary.

Also, use `instance` instead of `instances` in the path. There should only be one!