Initial implementation of instance specs (#146)
Instance specs are described more fully in RFD 283, but to summarize: An instance spec enumerates a Propolis instance's mainboard, devices, and backends and describes how these components are presented to the guest. Comparing two specs allows a migration source and migration target to agree on whether they have compatible device topologies--that is, whether the guest will observe some kind of unexpected or unsafe platform change if the source migrates to the target.

This PR:

- Defines the instance spec type in the propolis-client crate
- Adds a propolis-types crate to hold fundamental types that are shared amongst other Propolis crates
- Refactors propolis-server to initialize VMs in terms of their instance specs: instead of initializing devices and backends directly from REST API requests and config TOMLs, the server digests these inputs into a spec and initializes from the spec

Tested using cargo tests and ad hoc examination of Helios and Windows Server VMs running with these changes.