Allow instance_serial_task to block after last client disconnects (#160)
Change the serial task so that it passes websocket streams/sinks to 'for_each_concurrent' only if there's at least one such stream/sink. Otherwise, create a terminated fused future for these events. This prevents the task from specifying an automatically-completed future to select! when all serial console clients disconnect from a server, which can cause the server not to respond to new messages.

Tweak the select! order in the task so that incoming data from the intermediate buffer (the channel that receives messages from clients) is processed before new messages are read from the clients' streams. This allows the task to process connection-closed messages and to discard their streams before reading them again; done in the other order, the task will read the same 'closed' message multiple times, which can cause clients to be removed from the task's client list twice.