Add stat framework to Propolis (#147)
Add metrics to Propolis, and plumb through metrics from Crucible.

Creation of a metric server is determined at propolis-server start
time, though the actual metrics endpoint won't start until the
instance itself is started.  An optional metric_addr arg if provided
will indicate that a metrics endpoint will later be started.

On instance create, a ProducerRegistry can be created that can be
passed to libraries that wish to record stats. This allows Propolis
to not have to know what stats a library wishes to register.

A new struct, InstanceMetrics, has been added to Context. This is
used for tracking stats for an instance as well as containing the
ProducerRegistry that libraries can use to register with
Oximeter/Nexus for their own stats.

The CrucibleBackend Volume::construct() now takes a
ProducerRegistry that if present can be used to register metrics
with Oximeter.

Co-authored-by: Alan Hanson <alan@oxide.computer>