Have instance spec compat checks return `Result`s (#185)
Define two error types for use when comparing instance specs: the SpecMismatchDetails type contains enumerants that describe specific kinds of mismatches, and the MigrationCompatibilityError type wraps it with some additional context describing the fields that were being compared when an error was generated. Although most SpecMismatchDetails errors are specific enough to function as standalone errors, this scheme lets "generic" mismatches like "you had the wrong number of entries in a map" be returned alongside information about which specific map had the problem.