PHD: allow a single test case to launch multiple VMs (#186)
Add affordances for running multiple VMs in a single PHD test case. This largely involves making the VM factory hand out Dropshot/VNC server ports from a range instead of assuming that all VMs will use ports 9000/5900.

Add a multiple-VM smoke test. This test revealed a few tracing defects in the framework that this change aims to fix (e.g. some span usage bugs that made it hard to tell which serial console output came from which VM).

Finally, delete the uname smoke test, since it was originally just a proof-of-concept for the framework, and there's not much reason to spend time running it when other tests are online.