Derive vnic device names/backends from PCI path
The "name" field in a NetworkInterfaceRequest is the name of the host vNIC to
which the Propolis network device will bind. Because host vNIC names can change
over a migration, but device/backend names cannot, the former can't be used to
generate the latter.

Instead, generate device/backend names using PCI paths, which have to remain
stable over a migration.