Derive vnic device names/backends from PCI path (#204)
Two instance specs' network device and backend names must agree for the specs to be migration-compatible. Today, the spec generator derives these names from the name of the host vNIC to which the instance should bind, which name can and will change over a migration. Instead, generate network device/backend names from the NIC's PCI path, which is already not allowed to change during migration.