Slim down the live migration preamble (#221)
When transmitting an instance's specification in the LM preamble, send backends' names but not their configuration data. This reduces the cost to send the preamble, substantially simplifies instance spec versioning, and will (in a future PR) allow other large data (cloud-init bytes) to move back into the backend spec, simplifying the server code that has to deal with it.