PHD: add affordance for marking tests as skipped (#226)
Define a phd_skip! macro that allows a test to mark itself as skipped. This is useful in cases where a test case discovers that the runner was invoked in a configuration that prevents the test case from verifying anything meaningful. (For example, a test that wants to persist data on disk across a reboot may mark itself as skipped if the runner executes it with a guest image backed by a read-only file system.)