Split instance spec builder into generic and server-specific pieces (#225)
Refactor instance spec construction so that propolis-client has routines to incrementally build a spec and propolis-server contains only the bits needed to convert server request parameters into operations on the client's incremental builder. This sets up other clients (PHD in the short term, Nexus in the long term) with an incremental spec builder that they can use to construct specs to send to Propolis.