PHD: separate disk lifetimes from VM lifetimes (#228)
Rework PHD's guest disk management so that test VMs create temporary copies of guest images and boot from those instead of booting directly from the images in the artifact store. This obsoletes the framework's ZFS snapshot fixture, so remove it.

Also reorganize the VM configuration builder to accommodate this method of specifying disks. Along the way, make it more instance spec-driven to remove some getter routines and to make it easier to write functions that create migration target configs that refer to the same disks as their sources.

This creates a much more flexible interface for tests to create logical disks and attach them to VMs. This interface will feature in a future PR that will teach PHD to work with Crucible backends. It also significantly reduces CI execution times (by a factor of about 4 over previous runs).