server: Unify cleanup for external Stop request and guest-initiated shutdown.
We were only cleaning up the server-side resources if we got an
explicity external Stop request. But if the guest initiated a shutdown
we would correctly marked the VM as `Stopped` but never cleanup the
server's `ServiceProviders` nor ever mark it as `Destroyed`.

Instead this introduces a task that will trigger the cleanup when the VM
state worker exits.