server: bail early if the VNC server fails to start.
We were join'ing on the dropshot and VNC servers which meant we'd wait
for both futures to complete before bailing on an error. Came across
this when I started multiple instances but didn't specify different VNC
ports. The logs gave no indication otherwise.

Also pickup rfb change that makes `stop` not just signal the VNC server
to stop but also waits until it completes.