server: Wait for Instance to be dropped before setting api state to Destroyed.
Previously, there was a bit of a race condition where the completion of
the worker thread (and the subsequent dropping of WorkerStateInner)
would set the API instance state to Destroyed, but the actual Instance
itself might not've been destroyed yet. Some PHD tests would thus
sometimes leave left over VMM handles because the test framework
immediately killed the propolis-server process when it saw the instance
state get to Destroyed.