virtio early break, bump dlpi, no pub header-check
Break out of the guest packet processing loop early when there are no
bytes to read. This has a non-trivial impact on latency reduction.

Bump the dlpi dependency. The latest version is less chatty with logging
and the bug previously referenced is no longer an issue since we are not
using the async API exported by the dlpi create.

Set publish = false on the header check libraries. When importing
propolis libraries, because both viona-api and bhyve-api have a
sub-crate called header-check, cargo starts screaming.