Make the VM state driver unit testable (#281)
Refactor the VM controller so that the worker thread that accepts incoming commands (things like 'run', 'halt', or 'migrate') has well-defined interfaces to access VM controller state, to issue commands to vCPU tasks, and to interoperate with live migration tasks. This makes it possible to write test doubles for substantially all of the state worker's side-effect-producing entanglements, which in turn makes it much more feasible to unit test the state worker.