Replace migration transport layer with websockets
...and (or to be more precise, *because of*) update to Dropshot 0.9.x to
match Omicron main.

In Dropshot 0.9.x, an owned copy of the hyper::Request is no longer
available to endpoint handlers through the RequestContext. This poses a
problem for the custom wire protocol Propolis used for migration data
because we can no longer negotiate a HTTP upgrade.

So instead, let's switch out the custom data framing protocol for the
one provided in Dropshot `#[channel]` (that being the Websockets support
I added a while back) while keeping most of the data format the same.
In essence, this means just shoving it into a
`tungstenite::Message::Binary(vec![/* everything but the length, which
is now handled for us */])`.
The protocol version "negotiation" that was once done in the HTTP
upgrade header is now done by exchanging tungstenite::Message::Text(/*
the version string */) before handing the channel off.

I've also taken a couple small liberties as a matter of preference,
since we're breaking compatibility anyhow:
- the API endpoints are now
  `/instance/migrate/{migration_id}/[start|status]`, which I believe is
  more idiomatic REST than having the `migration_id` be in the query
  parameters of one and the request body of the other, especially when
  it's the *only* parameter in both cases.
- the `MessageType` tag byte in the websocket message is at the end of
  the message, as it means the `u64`s we read from the beginning of the
  `Vec` are more likely to be nicely aligned in memory, in case that
  ever matters (i.e. we someday end up running on some platform where
  that's necessary).