Notify the request queue of state driver work
Starting an instance allows it to migrate out and reboot, and stopping or
failing it causes requests to stop to be idempotently allowed and other requests
to fail. Add interfaces to the external request queue to allow the state driver
to notify it of these conditions.

Also fix some silly bugs (the inevitable "changed the queue but didn't signal
the condvar" bug; typo in the queue that checked the wrong flag when querying
if it's OK to be a migration source).

Finally, rearrange the state driver to be a bit more consistent about where it
calls out to other components vs. where it makes decisions about what actions
to take next.

Tested: cargo test, PHD.