Use the Failed terminal state; clean up migration in state driver
Move VMs to the Failed state when they're unable to start or when a live
migration in fails. Reorganize the `migrate_as_` functions in the state driver
to be a bit crisper about who is responsible for publishing the final migration
state (decide by fiat that it's supposed to be the migration tasks).