Remove MigrateStart instance state request (#332)
This request sets a flag in the VM controller that lets a VM act as a migration source if its state otherwise permits it. There's no way to clear the flag, which makes it tough to write an undo action for this in the Nexus live migration saga that (presumably) would issue this call.

Since the request has no other side effects, the VM controller already rejects other invalid requests to migrate out, and sled agent doesn't presently use this request type, just remove it entirely.

Tested with an ad hoc migration & PHD tests.