Use single channel for instance & migration state updates
N.B. This is a breaking change for users of the instance state monitor API
(such as the sled agent).

This allows state updates returned by the `instance_state_monitor` API contain
a single atomic snapshot of both an instance's state and its active migration
status.

This helps callers like sled agent that reason about instance and migration
states in tandem to do so without having to account for the fact that an
instance's state can change between the time a call to `monitor` returns and
the time a subsequent call to query migration status returns.

Tested: cargo test, PHD.