migration: require targets to receive final OK from sources
At the end of migration, require source and target to mutually affirm that
migration is completed, and make the source send the last "I'm OK" message in
the protocol.

This avoids situations where a target thinks migration has succeeded (because it
successfully sent a final "I'm OK" message to the source) but the source thinks
it has failed (because it never received that message).