Change state correctly after migration finishes; make use of the Failed state (#346)
Ensure that after a live migration in or out, the state driver transitions the VM to some sensible state. This was previously fixed for some cases by making migration targets go to the Failed state after failing to migrate in; this commit aims to finish the job for all migration outcomes.

Turn the Failed state into a terminal state that destroys the underlying VM so that the control plane doesn't have to coordinate requests to stop such a VM. Make the VM controller's Drop impl preserve the Failed state so that it's distinct from Destroyed (sled agent/Nexus treat Destroyed as Stopped, but they shouldn't do this for Failed).

Tweak the migration code in the state driver to be crisper about who's responsible for publishing migration state updates. Previously, the state driver published some of these; now it assumes the migration tasks always do.