migration: make source/target mutually acknowledge completion (#351)
In the current migration protocol's Finish phase, if the destination successfully sends its final "OK" message, it will think migration has succeeded irrespective of whether the source received or acknowledged that message. This can create a scenario where both source and target think they're supposed to be running a VM (the target because it thinks migration succeeded, the source because it thinks migration failed).

Fix this by changing this phase so that (a) the source succeeds only if it successfully sends a final "OK" message, and (b) the target succeeds only if it successfully receives that message. This creates the possibility that no one will run the VM (source sent the last handoff, target didn't receive it), but that's not much different from the VM successfully migrating and then immediately crashing (and is probably better than having two identical instances fighting over the same backends, etc.).

Also change this phase to publish the "Finish" state only after passing the point of no return. This lets callers use the Finish migration state to mean "migration completed successfully" without having to worry about a later transition to the Error state.