Support illumos#15143 data and vCPU pause/resume
With the integration of illumos#15143, additional pending interrupt and
exception state for guest vCPUs is exposed to userspace.  In order to
take a consistent snapshot of said state, as well as the guest devices
emulated in-kernel, propolis-server needs to issue VM_PAUSE requests
prior to exporting or importing state, and VM_RESUME requests when
starting a VM after a migration.

Importation of the "new" vCPU state is guarded with a version check for
now, so that propolis continues to work on hosts which have not yet
updated to bits featuring illumos#15143.