pull firmware builds from buildomat
Pick up a version of omicron-zone-package that allows package manifests to refer
to Buildomat artifacts. (Refer to the new bits by commit hash since a new
verison of omicron-zone-package hasn't been cut yet.) Download the zone image's
guest firmware from Buildomat instead of S3.

Use the image_debug build for now; this can easily be changed to image_release
later (just need to change the series and the digest in the package manifest).