Fix circular reference in async block scheduler (#381)
The async block scheduler holds a reference to its associated block device.
The device receives a closure that it invokes when it wants to notify the
scheduler that the device has work to pull. Today, the closure owns a strong
reference to the scheduler, creating a circular reference. Break the
cycle by making the closure refer to the scheduler's notification object
instead (similar to what the sync block scheduler does).

Tested by creating a VM with a Crucible disk, stopping it immediately, and
verifying that the underlying bhyve VMM disappears immediately as well.