migration: reduce page transfers during pause
Send all pages from the source to the target once before pausing, then pause,
then send only those pages that were modified after being sent the first time.
This drastically reduces the number of pages that have to be sent while the VM
is not running.

This change sends *all* pages in the pre-pause RAM transfer page, irrespective
of whether they're dirty. This is a low-subtlety way of handling the following
case:

1. Source reads its dirty page bitmap
2. Bhyve marks the previously-dirty pages as clean
3. Source sends the pages to the target
4. The target crashes and migration fails

To migrate correctly on a second attempt, the source needs to send the new
target all the pages it sent in step 2 (in addition to any other modified
pages). It'd be more efficient to remember exactly which pages need to be sent,
but for now, this change just sends all guest pages once, since the set of
required pages is guaranteed to be a subset of the set of all pages. Note that
it's still important to query the dirty bitmap in this phase (so that the pages
will only be transferred while paused if they were dirtied after being sent the
first time).

Tested: PHD incl. migration variations; 200 migrations of a guest running a
memory stress mix.