Bump MSRV to 1.67 and remove USDT cruft
It has been long enough since the required features for USDT to work
properly were stabilized in Rust.  The crate features to elide the probe
asm sites has been removed due to widespread support in the compiler.

This includes updates to p4rs and softnpu in order to clean up stale
dependencies to and older USDT.