Crucible tell Nexus when a scrub is done
When an instance is started with a volume that contains a read-only-parent,
crucible upstairs (after a slight delay) will now start scrubbing that
read-only-parent and copying the data to the underlying volume.  When that
scrub has completed, crucible will notify Nexus that the scrub is done and
it can remove that read-only-parent from the Volume.

Crucible has to know what the address the propolis server is running at, and
from that can determine the Nexus endpoint.