migration: implement protocol negotiation
Implement RFD 71-style protocol negotiation. On an attempt to migrate, the
destination sends the source a list of protocols of the form
"propolis-migrate-<encoding>/<version>". The source selects the highest-
versioned protocol that it understands and sends this protocol string back
to the destination. Thereafter the two sides use the agreed-upon protocol.