server: add RFD 71-style protocol negotiation (#405)
Negotiate migration protocols in the manner described in RFD 71. The
destination sends the source a list of protocols of the form
"propolis-migrate-<encoding>/<version>". The source selects from these the
highest-versioned protocol that it understands and sends its string back to
the destination. Thereafter the two sides use the agreed-upon protocol.

Supported protocols are enumerated in a new `migrate::protocol` module in
propolis-server. Protocol selections are plumbed to the `migrate::source`
and `migrate::destination` functions that create the migration runners. To
add a protocol version, these functions can either match on the selected
protocol to create a version-specific runner or pass the selected version
into the runner so that it can provide version-specific behavior.