only allow one request to reboot to be enqueued at a time
Change the VM controller request queue so that requesting to reboot an instance
causes subsequent requests to be ignored until a "Rebooted" state change arrives
from the state driver.

Beef up the PHD reboot test a little bit--it can't check for idempotency
explicitly because its Propolis may be able to reboot the VM as quickly as the
test can issue new reboot requests, but the modified test provides at least a
little bit of extra coverage. It also now checks that reboot requests are
rejected before an instance has started and after it has stopped.