PHD: add Ubuntu 22.04 guest adapter
This requires an Ubuntu 22.04 image that has been run through cloud-init such
that the user/password combination in the adapter can be used to log in via the
serial console. I've uploaded an image to catacomb that has this property.

Tested via a local run with the new guest OS type.