Add modules for components and V0 instance specs
Move "leaf" components of an instance spec into a `components` module and the
actual device spec, backend spec, and instance spec definitions into a new
`v0` module. No one refers to the new definitions yet; this is just a first step
toward getting everyone cut over to the new code.

The main difference between the old and new components is that all components
now have freestanding `struct`s instead of having a struct with fields common to
all devices and a "Kind" enum that specifies device-specific properties. That
is, instead of having a StorageDevice struct that has a StorageDeviceKind
member that specifies virtio- or NVMe-specific properties, there are now just
VirtioStorageDevices and NvmeStorageDevices. The goal of this change is to make
it easier to implement routines and traits on individual device spec types
without having to match on any enum variants.