Add migration compatibility trait module
Move the migration compatibility check traits into their own module and
implement them for components and v0 instance spec structures. Add copies of
the existing compatibility tests to the components/v0 modules.

With this change, all the existing instance spec functionality of old should
be replicated into the new modules. The next step is to start tearing out the
old ones and fixing the build.