Support for replacing a Crucible VolumeConstructionRequest
A new propolis-server API was added to support the replacement of
a downstairs via a new VolumeConstructionRequest.  To do this replacement,
we need the current VCR to go along with the new VCR, and we can get that
from the InstanceSpec.

As part of this change we have moved the VmObjects InstanceSpec behind a
Mutex, as a successful replacement of a crucible backends VCR should be
followed by a replacement of the VCR in the instance spec.

Added a new vcr option to crucible-cli which can call this API

This also includes an updated Crucible, as we need the latest version
that does not require a Logger with calls to target replace.