Wire up bits for CPUID customization
This is a first cut at making CPUID information customizable, rather
than using the in-kernel vmm logic which applies limited transformations
to the data exposed by the host CPU.  Basic plumbing in propolis-lib is
provided, which is wired into propolis-standalone.  The same
configuration syntax is defined for propolis-server, but it is going
unused for now until some internal structure is hammered out.