Overhaul block device interfaces
This reshapes how block devices (frontends) interact with their attached
backends.  Rather than associating a backend with a device when it is
created, backend and device are attached after the fact, when each has
had an opportunity to be initialized.

Through its attachment handle, the device is able to notify the backend
of new requests, but also query sizing parameters (previously specified
during device construction), pause and unpause request retrieval, and
(in the future) perform other tasks such as cache mode alteration.

The backend has a corresponding attachment handle through which it
retrieves pending requests from the device -- behavior which is largely
unchanged from the original structure.

Rather than store device-specific information required to issue request
completions to the guest, the device emulation is expecting to use a
`Tracking` structure which will store the completion data to be
retrieved using an ID injected into the Request, which is passed back
with its result when processed by the backend.  This tracking structure
also implements several generic USDT probes for tracing block events,
rather than requiring the use of per-device probes.